+++ RELEASE_NOTES.md
@@ -1,5 +1,7 @@
 Release Notes
 =============
+## vNext
+* Web App: Ability to set Startup Command directly (useful for Linux deployments).
 
 ## 1.7.14
 * Identity: Update the list of all RBAC roles to latest.